requirements.txt
streamlit
sentence-transformers
scikit-learn
nltk
pypdf
docx2txt